+++ src/main/java/org/jabref/logic/bibtexkeypattern/BibtexKeyPatternPreferences.java
@@ -4,32 +4,57 @@ import org.jabref.model.bibtexkeypattern.GlobalBibtexKeyPattern;
 
 public class BibtexKeyPatternPreferences {
 
+    public enum KeySuffix {
+        ALWAYS,         // CiteKeyA, CiteKeyB, CiteKeyC ...
+        SECOND_WITH_A,  // CiteKey, CiteKeyA, CiteKeyB ...
+        SECOND_WITH_B   // CiteKey, CiteKeyB, CiteKeyC ...
+    }
+
+    private final boolean shouldAvoidOverwriteCiteKey;
+    private final boolean shouldWarnBeforeOverwriteCiteKey;
+    private final boolean shouldGenerateCiteKeysBeforeSaving;
+    private final KeySuffix keySuffix;
     private final String keyPatternRegex;
     private final String keyPatternReplacement;
-    private final boolean alwaysAddLetter;
-    private final boolean firstLetterA;
+    private final String unwantedCharacters;
     private final GlobalBibtexKeyPattern keyPattern;
-    private Character keywordDelimiter;
-    private boolean avoidOverwritingCiteKey;
-    private String unwantedCharacters;
+    private final Character keywordDelimiter;
 
-    public BibtexKeyPatternPreferences(String keyPatternRegex,
+    public BibtexKeyPatternPreferences(boolean shouldAvoidOverwriteCiteKey,
+                                       boolean shouldWarnBeforeOverwriteCiteKey,
+                                       boolean shouldGenerateCiteKeysBeforeSaving,
+                                       KeySuffix keySuffix,
+                                       String keyPatternRegex,
                                        String keyPatternReplacement,
-                                       boolean alwaysAddLetter,
-                                       boolean firstLetterA,
+                                       String unwantedCharacters,
                                        GlobalBibtexKeyPattern keyPattern,
-                                       Character keywordDelimiter,
-                                       boolean avoidOverwritingCiteKey,
-                                       String unwantedCharacters) {
+                                       Character keywordDelimiter) {
 
+        this.shouldAvoidOverwriteCiteKey = shouldAvoidOverwriteCiteKey;
+        this.shouldWarnBeforeOverwriteCiteKey = shouldWarnBeforeOverwriteCiteKey;
+        this.shouldGenerateCiteKeysBeforeSaving = shouldGenerateCiteKeysBeforeSaving;
+        this.keySuffix = keySuffix;
         this.keyPatternRegex = keyPatternRegex;
         this.keyPatternReplacement = keyPatternReplacement;
-        this.alwaysAddLetter = alwaysAddLetter;
-        this.firstLetterA = firstLetterA;
+        this.unwantedCharacters = unwantedCharacters;
         this.keyPattern = keyPattern;
         this.keywordDelimiter = keywordDelimiter;
-        this.avoidOverwritingCiteKey = avoidOverwritingCiteKey;
-        this.unwantedCharacters = unwantedCharacters;
+    }
+
+    public boolean shouldAvoidOverwriteCiteKey() {
+        return shouldAvoidOverwriteCiteKey;
+    }
+
+    public boolean shouldWarnBeforeOverwriteCiteKey() {
+        return shouldWarnBeforeOverwriteCiteKey;
+    }
+
+    public boolean shouldGenerateCiteKeysBeforeSaving() {
+        return shouldGenerateCiteKeysBeforeSaving;
+    }
+
+    public KeySuffix getKeySuffix() {
+        return keySuffix;
     }
 
     public String getKeyPatternRegex() {
@@ -38,14 +63,6 @@ public class BibtexKeyPatternPreferences {
 
     public String getKeyPatternReplacement() {
         return keyPatternReplacement;
-    }
-
-    public boolean isAlwaysAddLetter() {
-        return alwaysAddLetter;
-    }
-
-    public boolean isFirstLetterA() {
-        return firstLetterA;
     }
 
     public String getUnwantedCharacters() {
@@ -58,9 +75,5 @@ public class BibtexKeyPatternPreferences {
 
     public Character getKeywordDelimiter() {
         return keywordDelimiter;
-    }
-
-    public boolean avoidOverwritingCiteKey() {
-        return avoidOverwritingCiteKey;
     }
 }